working on modal!!
